List of OSC Commands
Added a list of all the OCS commands outside of the JSON format to allow folks to test them quicker & easier. I've included a list of all of the areas that commands can target so you can search through the list a bit easier! (the -/ is included so you can jump the the headers to avoid going through to every command that targets that area. The current format is /msgType/targetArea/commandMsg/X argumentType. The /X is only included in commands that multiPath was true for. The argumentType is listed as null for PPCMeterMessages as indicated in the original documentation provided. If the command had a verfied argumentType listed it's shown as verified_argumentType. Otherwise the argument type is listed based on the most likely type given the message type (i.e. PPPCSwitch is listed as bool, PPCString as string, PPCFader as float, etc.).