QLab 4.7 File with all OSC Commands
Created a QLab 4.7 file with template cues and usable OSC commands for every command available. All of the commands are setup to target channel 2 with a value of 1 for the argument if it's a bool or float, and a value of TEST for string commands. Should make testing all of the commands quick and easy, but there's over 2k of them, so might take a while to get through them all.